Major Refactoring (#21)
Now supports direct imglib2 structure

* fixes tile size issue

* new export builder

* improves metadata transfer

* remove persist false in batch command for output folder

* code cleanup and documentation

* support of plane metadata transfer

* adds a way to cancel exports

* changes command GUI, specify lzw compression for temp files

* supports 2d rai input

* solve warnings using intellij, uses system array copy in scaler

+++ src/main/java/ch/epfl/biop/kheops/CZTRange.java
@@ -96,7 +96,7 @@ public class CZTRange {
 	 * @return the total number of planes to expect from this range
 	 */
 	public long getTotalPlanes() {
-		return rangeC.size() * rangeZ.size() * rangeT.size();
+		return (long) rangeC.size() * (long) rangeZ.size() * (long) rangeT.size();
 	}
 
 	public static Builder builder() {
